Added tests for Hierarchy and minimised existing tests
